Restructured code.
