+++ src/main/java/com/example/demo/DemoApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EnableJpaRepositories(basePackages = {"com.example.demo.model.repository"})
+@EnableJpaRepositories(basePackages = {"com.example.demo.dao"})
 public class DemoApplication {
 
 	public static void main(String[] args) {
