+++ src/main/java/com/example/demo/controller/DemoController.java
@@ -1,8 +1,8 @@
 package com.example.demo.controller;
 
 import com.example.demo.exceptions.ResourceNotFoundException;
+import com.example.demo.model.EmployeeEntity;
 import com.example.demo.model.Hello;
-import com.example.demo.model.entity.EmployeeEntity;
 import com.example.demo.service.EmployeeService;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
