+++ src/main/java/com/example/demo/dao/IEmployeeRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.dao;
 
-import com.example.demo.model.entity.EmployeeEntity;
+import com.example.demo.model.EmployeeEntity;
 import org.json.simple.parser.ParseException;
 
 import java.io.IOException;
