+++ src/main/java/com/example/demo/model/EmployeeEntity.java
@@ -1,4 +1,4 @@
-package com.example.demo.model.entity;
+package com.example.demo.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
