+++ src/main/java/com/example/demo/model/Users.java
@@ -1,4 +1,4 @@
-package com.example.demo.model.entity;
+package com.example.demo.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
