+++ src/test/java/com/example/demo/provider/MockDataProvider.java
@@ -1,6 +1,7 @@
 package com.example.demo.provider;
 
-import com.example.demo.model.entity.EmployeeEntity;
+
+import com.example.demo.model.EmployeeEntity;
 
 import java.util.Arrays;
 import java.util.List;
